+++ sdk/search/search-documents/src/serviceModels.ts
@@ -86,7 +86,9 @@ import {
   UniqueTokenFilter,
   WordDelimiterTokenFilter,
   SearchIndexerKnowledgeStoreProjection,
-  VectorSearchAlgorithmConfiguration as BaseVectorSearchAlgorithmConfiguration,
+  SearchIndexerIndexProjectionSelector,
+  NativeBlobSoftDeleteDeletionDetectionPolicy,
+  VectorSearchProfile,
 } from "./generated/service/models";
 
 import { PagedAsyncIterableIterator } from "@azure/core-paging";
@@ -612,7 +614,8 @@ export type SearchIndexerSkill =
   | TextTranslationSkill
   | DocumentExtractionSkill
   | WebApiSkill
-  | AzureMachineLearningSkill;
+  | AzureMachineLearningSkill
+  | AzureOpenAIEmbeddingSkill;
 
 /**
  * Contains the possible cases for CognitiveServicesAccount.
@@ -1001,7 +1004,7 @@ export interface SimpleField {
    * The name of the vector search algorithm configuration that specifies the algorithm and
    * optional parameters for searching the vector field.
    */
-  vectorSearchConfiguration?: string;
+  vectorSearchProfile?: string;
 }
 
 export function isComplexField(field: SearchField): field is ComplexField {
@@ -1313,6 +1316,10 @@ export interface SearchIndexerSkillset {
    * Definition of additional projections to azure blob, table, or files, of enriched data.
    */
   knowledgeStore?: SearchIndexerKnowledgeStore;
+  /**
+   *  Definition of additional projections to secondary search index(es).
+   */
+  indexProjections?: SearchIndexerIndexProjections;
   /**
    * The ETag of the skillset.
    */
@@ -2018,7 +2025,9 @@ export type SearchIndexerDataIdentity =
 /**
  * Contains the possible cases for DataDeletionDetectionPolicy.
  */
-export type DataDeletionDetectionPolicy = SoftDeleteColumnDeletionDetectionPolicy;
+export type DataDeletionDetectionPolicy =
+  | SoftDeleteColumnDeletionDetectionPolicy
+  | NativeBlobSoftDeleteDeletionDetectionPolicy;
 
 /**
  * Represents a datasource definition, which can be used to configure an indexer.
@@ -2077,19 +2086,29 @@ export interface SearchIndexerDataSourceConnection {
   encryptionKey?: SearchResourceEncryptionKey;
 }
 
-/**
- * Contains configuration options related to vector search.
- */
+/** Contains configuration options related to vector search. */
 export interface VectorSearch {
-  /**
-   * Contains configuration options specific to the algorithm used during indexing time.
-   */
-  algorithmConfigurations?: VectorSearchAlgorithmConfiguration[];
+  /** Defines combinations of configurations to use with vector search. */
+  profiles?: VectorSearchProfile[];
+  /** Contains configuration options specific to the algorithm used during indexing and/or querying. */
+  algorithms?: VectorSearchAlgorithmConfiguration[];
+  /** Contains configuration options on how to vectorize text vector queries. */
+  vectorizers?: VectorSearchVectorizer[];
 }
 
+/** Contains configuration options specific to the algorithm used during indexing and/or querying. */
 export type VectorSearchAlgorithmConfiguration =
   | BaseVectorSearchAlgorithmConfiguration
-  | HnswVectorSearchAlgorithmConfiguration;
+  | HnswVectorSearchAlgorithmConfiguration
+  | ExhaustiveKnnVectorSearchAlgorithmConfiguration;
+
+/** Contains configuration options specific to the algorithm used during indexing and/or querying. */
+export interface BaseVectorSearchAlgorithmConfiguration {
+  /** Polymorphic discriminator, which specifies the different types this object can be */
+  kind: VectorSearchAlgorithmKind;
+  /** The name to associate with this particular configuration. */
+  name: string;
+}
 
 /**
  * Contains configuration options specific to the hnsw approximate nearest neighbors algorithm
@@ -2135,5 +2154,120 @@ export interface HnswParameters {
    */
   metric?: VectorSearchAlgorithmMetric;
 }
+
+/** Contains configuration options specific to the exhaustive KNN algorithm used during querying, which will perform brute-force search across the entire vector index. */
+export type ExhaustiveKnnVectorSearchAlgorithmConfiguration =
+  BaseVectorSearchAlgorithmConfiguration & {
+    /** Polymorphic discriminator, which specifies the different types this object can be */
+    kind: "exhaustiveKnn";
+    /** Contains the parameters specific to exhaustive KNN algorithm. */
+    parameters?: ExhaustiveKnnParameters;
+  };
+
+/** Contains the parameters specific to exhaustive KNN algorithm. */
+export interface ExhaustiveKnnParameters {
+  /** The similarity metric to use for vector comparisons. */
+  metric?: VectorSearchAlgorithmMetric;
+}
+
+/** A dictionary of index projection-specific configuration properties. Each name is the name of a specific property. Each value must be of a primitive type. */
+export interface SearchIndexerIndexProjectionsParameters {
+  /** Describes unknown properties. The value of an unknown property can be of "any" type. */
+  [property: string]: any;
+  /** Defines behavior of the index projections in relation to the rest of the indexer. */
+  projectionMode?: IndexProjectionMode;
+}
+
+/** Definition of additional projections to secondary search indexes. */
+export interface SearchIndexerIndexProjections {
+  /** A list of projections to be performed to secondary search indexes. */
+  selectors: SearchIndexerIndexProjectionSelector[];
+  /** A dictionary of index projection-specific configuration properties. Each name is the name of a specific property. Each value must be of a primitive type. */
+  parameters?: SearchIndexerIndexProjectionsParameters;
+}
+
+/** Contains specific details for a vectorization method to be used during query time. */
+export interface BaseVectorSearchVectorizer {
+  /** Polymorphic discriminator, which specifies the different types this object can be */
+  kind: VectorSearchVectorizerKind;
+  /** The name to associate with this particular vectorization method. */
+  name: string;
+}
+
+/** Contains the parameters specific to using an Azure Open AI service for vectorization at query time. */
+export type AzureOpenAIVectorizer = BaseVectorSearchVectorizer & {
+  /** Polymorphic discriminator, which specifies the different types this object can be */
+  kind: "azureOpenAI";
+  /** Contains the parameters specific to Azure Open AI embedding vectorization. */
+  azureOpenAIParameters?: AzureOpenAIParameters;
+};
+
+/** Contains the parameters specific to generating vector embeddings via a custom endpoint. */
+export type CustomVectorizer = BaseVectorSearchVectorizer & {
+  /** Polymorphic discriminator, which specifies the different types this object can be */
+  kind: "customWebApi";
+  /** Contains the parameters specific to generating vector embeddings via a custom endpoint. */
+  customVectorizerParameters?: CustomVectorizerParameters;
+};
+
+/** Contains the parameters specific to generating vector embeddings via a custom endpoint. */
+export interface CustomVectorizerParameters {
+  /** The uri for the Web API. */
+  uri?: string;
+  /** The headers required to make the http request. */
+  httpHeaders?: { [propertyName: string]: string };
+  /** The method for the http request. */
+  httpMethod?: string;
+  /** The desired timeout for the request. Default is 30 seconds. */
+  timeout?: string;
+  /** Applies to custom endpoints that connect to external code in an Azure function or some other application that provides the transformations. This value should be the application ID created for the function or app when it was registered with Azure Active Directory. When specified, the vectorization connects to the function or app using a managed ID (either system or user-assigned) of the search service and the access token of the function or app, using this value as the resource id for creating the scope of the access token. */
+  authResourceId?: string;
+  /** The user-assigned managed identity used for outbound connections. If an authResourceId is provided and it's not specified, the system-assigned managed identity is used. On updates to the indexer, if the identity is unspecified, the value remains unchanged. If set to "none", the value of this property is cleared. */
+  authIdentity?: SearchIndexerDataIdentity;
+}
+
+/** Contains configuration options on how to vectorize text vector queries. */
+export type VectorSearchVectorizer =
+  | BaseVectorSearchVectorizer
+  | AzureOpenAIVectorizer
+  | CustomVectorizer;
+
+/** Contains the parameters specific to using an Azure Open AI service for vectorization at query time. */
+export interface AzureOpenAIParameters {
+  /** The resource uri for your Azure Open AI resource. */
+  resourceUri?: string;
+  /** ID of your Azure Open AI model deployment on the designated resource. */
+  deploymentId?: string;
+  /** API key for the designated Azure Open AI resource. */
+  apiKey?: string;
+  /** The user-assigned managed identity used for outbound connections. */
+  authIdentity?: SearchIndexerDataIdentity;
+}
+
+/** Allows you to generate a vector embedding for a given text input using the Azure Open AI service. */
+export interface AzureOpenAIEmbeddingSkill extends BaseSearchIndexerSkill {
+  /** Polymorphic discriminator, which specifies the different types this object can be */
+  odatatype: "#Microsoft.Skills.Text.AzureOpenAIEmbeddingSkill";
+  /** The resource uri for your Azure Open AI resource. */
+  resourceUri?: string;
+  /** ID of your Azure Open AI model deployment on the designated resource. */
+  deploymentId?: string;
+  /** API key for the designated Azure Open AI resource. */
+  apiKey?: string;
+  /** The user-assigned managed identity used for outbound connections. */
+  authIdentity?: SearchIndexerDataIdentity;
+}
+
+/** The similarity metric to use for vector comparisons. */
 export type VectorSearchAlgorithmMetric = "cosine" | "euclidean" | "dotProduct";
+
+export type VectorSearchAlgorithmKind = "hnsw" | "exhaustiveKnn";
+
+/**
+ * Defines behavior of the index projections in relation to the rest of the indexer.
+ */
+export type IndexProjectionMode = "skipIndexingParentDocuments" | "includeIndexingParentDocuments";
+
+export type VectorSearchVectorizerKind = "azureOpenAI" | "customWebApi";
+
 // END manually modified generated interfaces